# Optimization of data structure