--- a/woltka/align.py
+++ b/woltka/align.py
@@ -78,7 +78,7 @@ def plain_mapper(fh, fmt=None, n=1000000):
 
         # add subject to subject set of the same query Id
         if query == this:
-            subque[-1].add(subject)
+            subque[-1].append(subject)
 
         # when query Id changes,
         else:
@@ -101,7 +101,7 @@ def plain_mapper(fh, fmt=None, n=1000000):
 
             # create new query and subject set pair
             qry_append(query)
-            sub_append({subject})
+            sub_append([subject])
 
             # update current query Id
             this = query
@@ -314,6 +314,12 @@ def cigar_to_lens(cigar):
         Alignment length.
     int
         Offset in subject sequence.
+
+    Notes
+    -----
+    This function significantly benefits from LRU cache because high-frequency
+    CIGAR strings (e.g., "150M") are common and redundant calculations can be
+    saved.
     """
     align, offset = 0, 0
     n = ''  # current step size