… woltka/tests/data/README.md
… woltka/tests/data/align/burst/split/S01.map.bz2
… woltka/tests/data/align/burst/split/S02.map.bz2
… woltka/tests/data/align/burst/split/S03.map.bz2
… woltka/tests/data/align/burst/split/S04.map.bz2
… woltka/tests/data/align/burst/split/S05.map.bz2
… woltka/tests/data/output/bowtie2.free.tsv
… woltka/tests/data/output/burst.genus.map/S01.txt.gz
… woltka/tests/data/output/burst.genus.map/S02.txt.gz
… woltka/tests/data/output/burst.genus.map/S03.txt.gz
… woltka/tests/data/output/burst.genus.map/S04.txt.gz
… woltka/tests/data/output/burst.genus.map/S05.txt.gz
… woltka/tests/data/output/burst.genus.process.tsv
… woltka/tests/data/output/split.genus.tsv
… woltka/tests/data/output/split.process.tsv
… woltka/tests/test_align.py
… woltka/tests/test_cli.py
… woltka/tests/test_ordinal.py
… woltka/tests/test_workflow.py