The PR contains a series of optimizations.