@qiyunzhu 2020-06-24
Commit 9a00caf:

In `align.parse_align_file`, replace dict with two deques. See #54 for detailed rationales and benchmarks.

`align.parse_align_file`: 326 ms ± 765 µs => 310 ms ± 826 µs
`workflow.classify`: 392 ms +- 1.3 ms => 367 ms +- 1.06 ms

There is ~5% increase of performance.