@qiyunzhu 2020-06-24
Fixed a bug which would carry over the last element of last sample into the current sample in rare scenarios. This bug does not affect the current code, and this part will be re-implemented anyway. But a fix is still favorable.