@qiyunzhu 2020-06-26
Added a new test data set, which are BURST alignments against genes (not genomes). This dataset enables tests on strip-index and nucleotide-based analyses. Also added unit tests for it. Also enriched documentation.