@qiyunzhu 2020-06-28
#62 :  Used LRU cache which significantly improves performance.
#63 : Converted subjects into frozensets, so that they are hashable and can be queried in LRU cache.